% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigned-data.R
\docType{data}
\name{assigned_data}
\alias{assigned_data}
\title{Example analysis data}
\format{
S4 object of class AnalysisData from the metabolyseR package.
}
\source{
Example analysis data based on the Brachypodium distachyon
ecotype comparison FIE-MS metabolomic fingerprinting data set available
in the metaboData package. The features include putative molecular formula
assignments.
}
\usage{
assigned_data
}
\description{
Example analysis data
}
\keyword{datasets}
