% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{availableMethods}
\alias{availableMethods}
\title{Available functional enrichment methods}
\usage{
availableMethods()
}
\value{
A character vector of available methods.
}
\description{
Methods available for functional enrichment using the FELLA package
}
\examples{
availableMethods()
}
