% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentParameters.R
\name{enrichmentParameters}
\alias{enrichmentParameters}
\title{enrichmentParameters}
\usage{
enrichmentParameters(organism)
}
\arguments{
\item{organism}{KEGG organism ID}
}
\description{
Enrichment parameters.
}
