% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{example_analysis}
\alias{example_analysis}
\title{Example analysis data}
\format{
S4 object of class Analysis from the metabolyseR package.
}
\source{
Example analysis data based on the Brachypodium distachyon 
ecotype comparison FIE-MS metabolomic fingerprinting data set available 
in the metaboData package.
}
\usage{
example_analysis
}
\description{
Example analysis data
}
\keyword{datasets}
