% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{example_assignment}
\alias{example_assignment}
\title{Example molecular formula assignment data}
\format{
S4 object of class Assignment from the MFassign package.
}
\source{
Example analysis data based on the Brachypodium distachyon
ecotype comparison FIE-MS metabolomic fingerprinting data set available 
in the metaboData package.
}
\usage{
example_assignment
}
\description{
Example molecular formula assignment data
}
\keyword{datasets}
