% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{functionalEnrichment}
\alias{functionalEnrichment}
\alias{functionalEnrichment,RandomForest-method}
\title{functional enrichment}
\usage{
functionalEnrichment(
  x,
  organism,
  methods = availableMethods(),
  adduct_rules_table = mzAnnotation::adduct_rules(),
  organism_graph = organismGraph(organism)
)

\S4method{functionalEnrichment}{RandomForest}(
  x,
  organism,
  methods = availableMethods(),
  adduct_rules_table = mzAnnotation::adduct_rules(),
  organism_graph = organismGraph(organism)
)
}
\arguments{
\item{x}{object of S4 class RandomForest}
}
\description{
Functional enrichment.
}
\examples{
\dontrun{
## Generate enrichment parameters
parameters <- enrichmentParameters('bdi')

## Select only "diffusion" enrichment
functional(parameters) <- list(methods = 'diffusion')

## Run functional enrichment
fe <- functionalEnrichment(example_analysis,example_assignment,parameters)
}
}
