% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism-data.R
\name{organismData}
\alias{organismData}
\title{Organism data}
\usage{
organismData(
  organism,
  methods = availableMethods(),
  filter_path = NULL,
  database_directory = organism,
  internal_directory = TRUE,
  damping_factor = 0.85,
  niter = 100
)
}
\arguments{
\item{organism}{the KEGG code for the organism of interest}

\item{methods}{the enrichment techniques to build. Any returned by \code{availableMethods}.}

\item{filter_path}{argument to pass to argument \code{filter.path} of \code{FELLA::buildGraphFromKEGGREST}. A vector of regular expressions to match pathways to exclude.}

\item{database_directory}{argument to pass to argument \code{databaseDir} of \code{FELLA::buildDataFromGraph} and \code{FELLA::loadKEGGdata}. The directory name/path in which to save the KEGG data.}

\item{internal_directory}{logical. Argument to pass to argument \code{internalDir} of \code{FELLA::buildDataFromGraph} and \code{FELLA::loadKEGGdata}. Should the save directory be internal to the FELLA package directory?}

\item{damping_factor}{argument to pass to argument \code{dampingFactor} of \code{FELLA::buildDataFromGraph}. A value between 0 and 1 for the PageRank damping factor.}

\item{niter}{argument to pass to argument \code{niter} of \code{FELLA::buildDataFromGraph}. A value between 10 and 1000. The number of iterations to estimate the values for CC size.}
}
\value{
An object of S4 class \code{FELLA.DATA} containing the KEGG data for the specified organism.
}
\description{
Load or build the organism specific KEGG datafor enrichment analysis.
This provides a convenience wrapper around \code{FELLA::loadKEGGdata}, \code{FELLA::buildGraphFromKEGGREST} and
\code{FELLA::buildDataFromGraph}. See the documentation of these functions for further information.
}
\examples{
## Load the example organism data available from within the package
organismData(
  'bdi',
  database_directory = system.file(
    'bdi',
    package = 'riches'),
  internal_directory = FALSE
)
}
