% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism-graph.R
\name{organismGraph}
\alias{organismGraph}
\title{Organism graph}
\usage{
organismGraph(organism = "hsa", filter.path = NULL, internalDir = TRUE)
}
\arguments{
\item{organism}{KEGG organism code}
}
\description{
Gather organism specific KEGG graph ready for enrichment analysis
}
