% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organismNetwork.R
\name{organismNetwork}
\alias{organismNetwork}
\title{organismNetwork}
\usage{
organismNetwork(organism = "hsa")
}
\arguments{
\item{organism}{KEGG organism code}
}
\description{
Gather organism specific KEGG network ready for enrichment analysis
}
