% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-enrichment.R
\name{structuralClassifications}
\alias{structuralClassifications}
\alias{structuralClassifications,StructuralEnrichment-method}
\alias{explanatoryFeatures,StructuralEnrichment-method}
\alias{enrichmentResults,StructuralEnrichment-method}
\title{StructuralEnrichment S4 class accessors}
\usage{
structuralClassifications(x)

\S4method{structuralClassifications}{StructuralEnrichment}(x)

\S4method{explanatoryFeatures}{StructuralEnrichment}(x)

\S4method{enrichmentResults}{StructuralEnrichment}(x)
}
\arguments{
\item{x}{object of S4 class \code{StructuralEnrichment}}
}
\value{
A tibble containing the accessed information.
}
\description{
Accessor methods for the \code{StructuralEnrichment} S4 class.
}
\examples{
## Perform random forest on the example data 
random_forest <- assigned_data \%>\% 
  metabolyseR::randomForest(
    cls = 'class'
  )

## Perform functional enrichment analysis
enrichment_results <- structuralEnrichment(
  random_forest,
  structural_classifications
)

## The m/z feature structural classifcations
structuralClassifications(enrichment_results)

## Access the explanatory features used for structural enrichment analysis
explanatoryFeatures(enrichment_results)

## Access the structral enrichment analysis results
enrichmentResults(enrichment_results)
}
