% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-classifications.R
\docType{data}
\name{structural_classifications}
\alias{structural_classifications}
\title{Example structural classifications}
\format{
A tibble containing consensus structural classifications for the \emph{m/z} features found
in the \code{assigned_data} example data set.
}
\source{
These example consensus structural classifications were generated using the
\href{https://jasenfinch.github.io/construction/}{{construction}} package (v0.3.0) and are based on
the example analysis data (\code{assigned_data}) based on the \emph{Brachypodium distachyon} ecotype
comparison FIE-MS metabolomic fingerprinting data set available in the
\href{https://aberhrml.github.io/metaboData/}{{metaboData}} package.
}
\usage{
structural_classifications
}
\description{
Example structural classifications
}
\keyword{datasets}
