% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-enrichment.R
\name{structuralEnrichment}
\alias{structuralEnrichment}
\alias{structuralEnrichment,RandomForest,tbl_df-method}
\title{Structural enrichment}
\usage{
structuralEnrichment(
  x,
  structural_classifications,
  p_adjust_method = "bonferroni",
  ...
)

\S4method{structuralEnrichment}{RandomForest,tbl_df}(
  x,
  structural_classifications,
  p_adjust_method = "bonferroni",
  ...
)
}
\arguments{
\item{x}{an object of S4 class \code{RandomForest}}

\item{structural_classifications}{the structral classifications corresponding to the \emph{m/z} features present in the object specified for argument \code{x}. This should either be a tibble as returned by \code{construction::classifications()} or an object of S4 class \code{Construction}.}

\item{p_adjust_method}{the p-value adjustment method. One of those returned from \code{p.adjust.methods}.}

\item{...}{arguments to pass to \code{metabolyseR::explanatoryFeatures()}}
}
\value{
An object of S4 class \code{StructuralEnrichment}.
}
\description{
Perform structural enrichment using over-representation analysis of explanatory \emph{m/z} features from random forest.
}
\details{
Over-representation analysis is performed on the explanatory \emph{m/z} features for each structural class within each experimental class comparison using the Fisher's Exact Test.
}
\examples{
## Perform random forest on the example data 
random_forest <- assigned_data \%>\% 
  metabolyseR::randomForest(
    cls = 'class'
  )

## Perform structural enrichment analysis using the example structural classifications
structuralEnrichment(
  random_forest,
  structural_classifications
)
}
