% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{hits}
\alias{hits}
\alias{hits,FunctionalEnrichment-method}
\alias{explanatoryFeatures,FunctionalEnrichment-method}
\alias{enrichmentResults}
\alias{enrichmentResults,FunctionalEnrichment-method}
\title{FunctionalEnrichment S4 class accessors}
\usage{
hits(x)

\S4method{hits}{FunctionalEnrichment}(x)

\S4method{explanatoryFeatures}{FunctionalEnrichment}(x)

enrichmentResults(x)

\S4method{enrichmentResults}{FunctionalEnrichment}(x)
}
\arguments{
\item{x}{object of S4 class \code{FunctionalEnrichment}}
}
\value{
A tibble or a list of objects of \code{FELLA.USER} S4 class depending on the method used.
}
\description{
Accessor methods for the \code{FunctionalEnrichment} S4 class.
}
\examples{
## Perform random forest on the example data 
random_forest <- assigned_data \%>\% 
metabolyseR::randomForest(
  cls = 'class'
)

## Perform functional enrichment analysis
enrichment_results <- functionalEnrichment(
  random_forest,
  'bdi',
  methods = 'hypergeom',
  organism_data = organismData(
    'bdi',
    database_directory = system.file(
      'bdi',
      package = 'riches'),
    internal_directory = FALSE
  )
)

## Access the m/z feature KEGG compound matches
hits(enrichment_results)

## Access the explanatory features used for functional enrichment
explanatoryFeatures(enrichment_results)

## Access the functional enrichment results
enrichmentResults(enrichment_results)
}
