% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{functional}
\alias{functional}
\alias{functional<-}
\alias{functional,EnrichmentParameters-method}
\alias{functional<-,EnrichmentParameters-method}
\title{Get and set enrichment parameters}
\usage{
functional(x)

functional(x) <- value

\S4method{functional}{EnrichmentParameters}(x)

\S4method{functional}{EnrichmentParameters}(x) <- value
}
\arguments{
\item{value}{value to set}

\item{S4}{object of class EnrichmentParameters}
}
\description{
Retrieve or set enrichment parameters.
}
