% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{functionalEnrichment}
\alias{functionalEnrichment}
\alias{functionalEnrichment,RandomForest-method}
\title{functional enrichment}
\usage{
functionalEnrichment(
  x,
  organism,
  methods = availableMethods(),
  organism_data = organismData(organism),
  adduct_rules_table = adduct_rules(),
  ...
)

\S4method{functionalEnrichment}{RandomForest}(
  x,
  organism,
  methods = availableMethods(),
  organism_data = organismData(organism),
  adduct_rules_table = adduct_rules(),
  ...
)
}
\arguments{
\item{x}{object of S4 class \code{RandomForest}}

\item{organism}{the KEGG code for the organism of interest}

\item{methods}{the enrichment techniques to build. Any returned by \code{availableMethods}.}

\item{organism_data}{an object of S4 class \code{FELLA.DATA}}

\item{adduct_rules_table}{the adduct ionisation rules for matching m/z features to KEGG compounds. Format should be as returned from \code{mzAnnotation::adduct_rules}.}

\item{...}{arguments to pass to \code{metabolyseR::explanatoryFeatures}}
}
\description{
Functional enrichment.
}
\examples{
## Perform random forest on the example data 
random_forest <- assigned_data \%>\% 
metabolyseR::randomForest(
  cls = 'class'
)

## Perform functional enrichment analysis
functionalEnrichment(
  random_forest,
  'bdi',
  methods = 'hypergeom',
  organism_data = organismData(
    'bdi',
    database_directory = system.file(
      'bdi',
      package = 'riches'),
    internal_directory = FALSE
  )
)
}
