% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/functionalEnrichment.R
\name{functionalEnrichment}
\alias{functionalEnrichment}
\alias{functionalEnrichment,Analysis,Assignment,EnrichmentParameters-method}
\title{functionalEnrichment}
\usage{
functionalEnrichment(analysis, assignment, parameters)

\S4method{functionalEnrichment}{Analysis,Assignment,EnrichmentParameters}(analysis, assignment, parameters)
}
\arguments{
\item{analysis}{S4 object of class Analysis}

\item{assignment}{S4 object of class Assignment}

\item{parameters}{S4 object of class EnrichmentParameters}
}
\description{
Functional enrichment.
}
\examples{
\dontrun{
## Generate enrichment parameters
parameters <- enrichmentParameters('bdi')

## Select only "diffusion" enrichment
functional(parameters) <- list(methods = 'diffusion')

## Run functional enrichment
fe <- functionalEnrichment(example_analysis,example_assignment,parameters)
}
}
