% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-enrichment.R
\name{structuralEnrichment}
\alias{structuralEnrichment}
\alias{structuralEnrichment,RandomForest,tbl_df-method}
\alias{structuralEnrichment,RandomForest,Construction-method}
\title{Structural enrichment}
\usage{
structuralEnrichment(
  x,
  structural_classifications,
  p_adjust_method = "bonferroni",
  split = c("none", "trends"),
  ...
)

\S4method{structuralEnrichment}{RandomForest,tbl_df}(
  x,
  structural_classifications,
  p_adjust_method = "bonferroni",
  split = c("none", "trends"),
  ...
)

\S4method{structuralEnrichment}{RandomForest,Construction}(
  x,
  structural_classifications,
  p_adjust_method = "bonferroni",
  split = c("none", "trends"),
  ...
)
}
\arguments{
\item{x}{an object of S4 class \code{RandomForest}}

\item{structural_classifications}{the structral classifications corresponding to the \emph{m/z} features present in the object specified for argument \code{x}. This should either be a tibble as returned by \code{construction::classifications()} or an object of S4 class \code{Construction}.}

\item{p_adjust_method}{the p-value adjustment method. One of those returned from \code{p.adjust.methods}.}

\item{split}{split the explanatory features into further groups based on their trends. See details.}

\item{...}{arguments to pass to \code{metabolyseR::explanatoryFeatures()}}
}
\value{
An object of S4 class \code{StructuralEnrichment}.
}
\description{
Perform structural enrichment using over-representation analysis of explanatory \emph{m/z} features from random forest.
}
\details{
Over-representation analysis is performed on the explanatory \emph{m/z} features for each structural class
within each experimental class comparison using the Fisher's Exact Test.

For argument \code{split = 'trends'}, the explanatory features can be split into further groups
based on their trends. This is not supported for unsupervised random forest.

For random forest classification, this is for binary comparisons only. Functional enrichment
is performed seperately on the up and down regulated explanatory features for each comparison. The
\verb{up regulated} and \verb{down regulated} groups are based on the trends of log2 ratios between
the comparison classes. \verb{up regulated} explanatory features have a higher median intensity
in the right-hand class compared to the left-hand class of the comparison. The opposite is true
for the \verb{down regulated} explanatory features.

For random forest regression, the explanatory features are split based on their Spearman's
correlation coefficient with the response variable prior to functional enrichment analysis
giving \verb{positively correlated} and \verb{negatively correlated} subgroups.
}
\examples{
## Perform random forest on the example data 
random_forest <- assigned_data \%>\% 
  metabolyseR::randomForest(
    cls = 'class'
  )

## Perform structural enrichment analysis using the example structural classifications
structuralEnrichment(
  random_forest,
  structural_classifications
)

## An example using split trends
## Perform random forest on the example data 
random_forest <- assigned_data \%>\% 
  metabolyseR::randomForest(
    cls = 'class',
    binary = TRUE
  )

## Perform structural enrichment analysis using the example structural classifications
structuralEnrichment(
  random_forest,
  structural_classifications,
  split = 'trends'
)
}
